==> scripts/03_data_analysis/Study 1/01_study1_data_cleaning.ipynb <==
#%pip install davos
import davos

smuggle pandas as pd
smuggle numpy as np
smuggle os

current_directory = os.getcwd()
new_directory = "/Users/suelim/Documents/Research/DONE/Evaluation_Vaping_Messages/github/Evaluation_Vaping_Messages/data/02_Data/study1"
os.chdir(new_directory)

df = pd.read_csv("./01_raw_data_anonymized/study1_prolific_anonymized.csv")

# Identify pairs of columns to combine based on the mapping you provided
pairs = [
    ("C_Message Rank_" + str(i), "T_Message Ranking_" + str(i)) for i in range(1, 31)
] + [
    ("C_Dis_AI" + str(i), "T_Dis_AI" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Dis_Hum" + str(i), "T_Dis_Hum" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Con_AI" + str(i), "T_Con_AI" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Con_Hum" + str(i), "T_Con_Hum" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Unp_AI" + str(i), "T_Unp_AI" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Unp_Hum" + str(i), "T_Unp_Hum" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Lapp_AI" + str(i), "T_Lapp_AI" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Lapp_Hum" + str(i), "T_Lapp_Hum" + str(i) + "_1") for i in range(1, 16)
]

# Create new columns without modifying the original DataFrame in a loop
new_cols = {}
for c_col, t_col in pairs:
    combined_col_name = c_col.replace('C_', '')
    new_cols[combined_col_name] = df[c_col].fillna(df[t_col])

# Concatenate the new columns to the original DataFrame
df = pd.concat([df, pd.DataFrame(new_cols)], axis=1)

df2 = df
cols2 = df2.columns.tolist()
cols2.insert(0, cols2.pop(cols2.index('Experimental_Group')))
df2 = df2[cols2]
df2['Experimental_Group'] = df2['Experimental_Group'].map({0: 'Not Disclosed', 1: 'Disclosed'})
df2["Duration (in seconds)"] = pd.to_numeric(df2["Duration (in seconds)"])
df2 = df2[(df2["Duration (in seconds)"] > 300) & (df2["Finished"] == 1)]
# Find the index positions of 'Finished' and 'Age'
Finished_idx = cols2.index('Finished')
age_idx = cols2.index('Age')
# Reorder the columns
new_cols_order = cols2[:Finished_idx + 1] + cols2[age_idx:] + cols2[Finished_idx + 1:age_idx]
# Reassign the columns to the DataFrame based on the new order
df2 = df2[new_cols_order]
df2['recruitment_platform'] = 'Prolific'
df2.head()

df2.to_csv('./02_cleaned_data_file/study1_prolific_cleaned.csv', index=False)

df_sona = pd.read_csv("./01_raw_data_anonymized/study1_sona_anonymized.csv")

cols_sona = df_sona.columns.tolist()

# Extract unique numbers from C_Message Rank_ columns
message_rank_numbers = sorted([int(col.split('_')[-1]) for col in cols_sona if col.startswith('C_Message Rank_')])

# Generate pairs based on actual columns in dataframe for C_Message Rank_ and T_Message Ranking_
message_pairs = [("C_Message Rank_" + str(i), "T_Message Ranking_" + str(i)) for i in message_rank_numbers]

# Other pairs remain the same
other_pairs = [
    ("C_Dis_AI_" + str(i), "T_Dis_AI_" + str(i)) for i in range(1, 16)
] + [
    ("C_Dis_hum_" + str(i), "T_Dis_hum_" + str(i)) for i in range(1, 16)
] + [
    ("C_Con_AI_" + str(i), "T_Con_AI_" + str(i)) for i in range(1, 16)
] + [
    ("C_Con_hum_" + str(i), "T_Con_hum_" + str(i)) for i in range(1, 16)
] + [
    ("C_Unp_AI_" + str(i), "T_Unp_AI_" + str(i)) for i in range(1, 16)
] + [
    ("C_Unp_hum_" + str(i), "T_Unp_hum_" + str(i)) for i in range(1, 16)
] + [
    ("C_Lapp_AI_" + str(i), "T_Lapp_AI_" + str(i)) for i in range(1, 16)
] + [
    ("C_Lapp_hum_" + str(i), "T_Lapp_hum_" + str(i)) for i in range(1, 16)
]

# Combine pairs
pairs2 = message_pairs + other_pairs

# Create new columns without modifying the original DataFrame in a loop
new_cols_sona = {}
for c_col, t_col in pairs2:
    combined_col_name = c_col.replace('C_', '')
    new_cols_sona[combined_col_name] = df_sona[c_col].fillna(df_sona[t_col])

# Concatenate the new columns to the original DataFrame
df_sona = pd.concat([df_sona, pd.DataFrame(new_cols_sona)], axis=1)

df_sona2 = df_sona
df_sona2['Experimental_Group'] = np.where(df_sona2['T_Message Ranking_1'].isna(), 'Not Disclosed', np.where(df_sona2['C_Message Rank_1'].isna(), 'Disclosed', np.nan))
cols_sona2 = df_sona2.columns.tolist()
cols_sona2.insert(0, cols_sona2.pop(cols_sona2.index('Experimental_Group')))
df_sona2 = df_sona2[cols_sona2]
df_sona2["Duration (in seconds)"] = pd.to_numeric(df_sona2["Duration (in seconds)"])
df_sona2 = df_sona2[(df_sona2["Duration (in seconds)"] > 300) & (df_sona2["Finished"] == 1)]
# Find the index positions of 'Finished' and 'Age'
sona_Finished_idx = cols_sona2.index('Finished')
sona_age_idx = cols_sona2.index('Age')
# Reorder the columns
new_cols_order2 = cols_sona2[:sona_Finished_idx + 1] + cols_sona2[sona_age_idx:] + cols_sona2[sona_Finished_idx + 1:sona_age_idx]
# Reassign the columns to the DataFrame based on the new order
df_sona2 = df_sona2[new_cols_order2]
df_sona2['recruitment_platform'] = 'SONA'
df_sona2.head()

df_sona2 = df_sona2.rename(columns=lambda x: x.replace('_hum', '_Hum') if '_hum' in x else x)

df_sona2.to_csv('./02_cleaned_data_file/study1_sona_cleaned.csv', index=False)

prolific = pd.read_csv("./02_cleaned_data_file/study1_prolific_cleaned.csv")

selected_columns = [0] + list(range(5, 157)) + [457]
prolific_cleaned = prolific.iloc[:, selected_columns]
prolific_cleaned.head()

prolific_cols = prolific_cleaned.columns.tolist()
prolific_cols

column_mapping = {
    f"Message Rank_{i}": f"Rank_AI{i}" if i <= 15 else f"Rank_Human{i-15}"
    for i in range(1, 31)
}

prolific_cleaned = prolific_cleaned.rename(columns=column_mapping)
prolific_cleaned.head()

sona = pd.read_csv("./02_cleaned_data_file/study1_sona_cleaned.csv")

sona_cleaned = sona.iloc[:, selected_columns]
sona_cleaned.head()

sona_cols = sona_cleaned.columns.tolist()
sona_cols

sona_cleaned.rename(columns={
    "Message Rank_1": "Rank_AI1",
    "Message Rank_4": "Rank_AI2",
    "Message Rank_5": "Rank_AI3",
    "Message Rank_6": "Rank_AI4",
    "Message Rank_7": "Rank_AI5",
    "Message Rank_9": "Rank_AI6",
    "Message Rank_11": "Rank_AI7",
    "Message Rank_12": "Rank_AI8",
    "Message Rank_13": "Rank_AI9",
    "Message Rank_14": "Rank_AI10",
    "Message Rank_15": "Rank_AI11",
    "Message Rank_16": "Rank_AI12",
    "Message Rank_17": "Rank_AI13",
    "Message Rank_18": "Rank_AI14",
    "Message Rank_19": "Rank_AI15",
    "Message Rank_20": "Rank_Human1",
    "Message Rank_21": "Rank_Human2",
    "Message Rank_22": "Rank_Human3",
    "Message Rank_35": "Rank_Human4",
    "Message Rank_24": "Rank_Human5",
    "Message Rank_25": "Rank_Human6",
    "Message Rank_26": "Rank_Human7",
    "Message Rank_27": "Rank_Human8",
    "Message Rank_28": "Rank_Human9",
    "Message Rank_29": "Rank_Human10",
    "Message Rank_30": "Rank_Human11",
    "Message Rank_31": "Rank_Human12",
    "Message Rank_32": "Rank_Human13",
    "Message Rank_23": "Rank_Human14",
    "Message Rank_36": "Rank_Human15"}, inplace=True)
sona_cleaned.columns = sona_cleaned.columns.str.replace(r'_(?=[0-9])', '', regex=True)
sona_cleaned.head()

combined_df = pd.concat([prolific_cleaned, sona_cleaned], ignore_index=True, sort=False)
combined_df['Participant'] = range(1, len(combined_df) + 1)
cols_comb = combined_df.columns.tolist()
cols_comb.insert(0, cols_comb.pop(cols_comb.index('Participant')))
combined_df = combined_df[cols_comb]
combined_df

combined_df.to_csv('./02_cleaned_data_file/study1_data_file.csv', index=False)
